refactor(domain): STTRepository와 STTPermissionRepository 통합
- STTRepository에 checkSTTPermission/requestSTTPermission 메서드 추가
- STTPermissionRepository 프로토콜 삭제
- CheckSTTPermissionUseCase·RequestSTTPermissionUseCase가 STTRepository 의존
- MockSTTPermissionRepository를 MockSTTRepository에 통합
- DefaultSTTRepository의 STTPermissionRepository conformance 제거
