+++ Data/Sources/Repositories/VoiceNotes/DefaultSTTRepository.swift
@@ -3,7 +3,7 @@ import Domain
 import Foundation
 
 /// 음성 인식(STT) 리포지토리 기본 구현체.
-public struct DefaultSTTRepository: STTRepository, STTPermissionRepository {
+public struct DefaultSTTRepository: STTRepository {
     private let service: any STTService
 
     public init(service: any STTService) {
