+++ Domain/Tests/Interfaces/Mocks/VoiceNote/MockSTTRepository.swift
@@ -4,20 +4,42 @@ import XCTest
 
 actor MockSTTRepository: STTRepository {
     private var result: Result<Transcript, STTRepositoryError>?
+    private var checkResult: Result<PermissionStatus, STTPermissionRepositoryError>?
+    private var requestResult: Result<PermissionStatus, STTPermissionRepositoryError>?
 
     private var actualCallCount = 0
     private var actualAudioFileURL: URL?
+    private var actualCheckSTTPermissionCallCount = 0
+    private var actualRequestSTTPermissionCallCount = 0
 
     private var expectedCallCount: Int?
     private var expectedAudioFileURL: URL?
+    private var expectedCheckSTTPermissionCallCount: Int?
+    private var expectedRequestSTTPermissionCallCount: Int?
 
     func setResult(_ result: Result<Transcript, STTRepositoryError>) {
         self.result = result
     }
 
+    func setCheckResult(_ result: Result<PermissionStatus, STTPermissionRepositoryError>) {
+        checkResult = result
+    }
+
+    func setRequestResult(_ result: Result<PermissionStatus, STTPermissionRepositoryError>) {
+        requestResult = result
+    }
+
     func expectTranscribe(callCount: Int, audioFileURL: URL? = nil) {
         expectedCallCount = callCount
         expectedAudioFileURL = audioFileURL
+    }
+
+    func expectCheckSTTPermission(callCount: Int) {
+        expectedCheckSTTPermissionCallCount = callCount
+    }
+
+    func expectRequestSTTPermission(callCount: Int) {
+        expectedRequestSTTPermissionCallCount = callCount
     }
 
     func verify(file: StaticString = #filePath, line: UInt = #line) {
@@ -29,6 +51,24 @@ actor MockSTTRepository: STTRepository {
         if let expectedURL = expectedAudioFileURL {
             XCTAssertEqual(
                 actualAudioFileURL, expectedURL, "변환 오디오 파일 URL이 일치하지 않습니다.", file: file, line: line
+            )
+        }
+        if let expected = expectedCheckSTTPermissionCallCount {
+            XCTAssertEqual(
+                actualCheckSTTPermissionCallCount,
+                expected,
+                "STT 권한 확인 호출 횟수가 일치하지 않습니다.",
+                file: file,
+                line: line
+            )
+        }
+        if let expected = expectedRequestSTTPermissionCallCount {
+            XCTAssertEqual(
+                actualRequestSTTPermissionCallCount,
+                expected,
+                "STT 권한 요청 호출 횟수가 일치하지 않습니다.",
+                file: file,
+                line: line
             )
         }
     }
@@ -47,6 +87,34 @@ actor MockSTTRepository: STTRepository {
             throw .unknown(
                 NSError(domain: "MockSTTRepository.result", code: -1)
             )
+        }
+    }
+
+    func checkSTTPermission() async throws(STTPermissionRepositoryError) -> PermissionStatus {
+        actualCheckSTTPermissionCallCount += 1
+
+        switch checkResult {
+        case .success(let state):
+            return state
+        case .failure(let error):
+            throw error
+        case .none:
+            XCTFail("MockSTTRepository.checkResult 가 설정되지 않았습니다.")
+            throw .unknown(NSError(domain: "MockSTTRepository.checkResult", code: -1))
+        }
+    }
+
+    func requestSTTPermission() async throws(STTPermissionRepositoryError) -> PermissionStatus {
+        actualRequestSTTPermissionCallCount += 1
+
+        switch requestResult {
+        case .success(let state):
+            return state
+        case .failure(let error):
+            throw error
+        case .none:
+            XCTFail("MockSTTRepository.requestResult 가 설정되지 않았습니다.")
+            throw .unknown(NSError(domain: "MockSTTRepository.requestResult", code: -1))
         }
     }
 }
